# api/docs/swagger.yaml
basePath: /
definitions:
  schemas.AccountSchema:
    properties:
      amount:
        $ref: '#/definitions/schemas.AmountSchema'
      colorHex:
        example: '#1E88E5'
        type: string
      description:
        example: Main salary card
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Monobank card
        type: string
    type: object
  schemas.AmountSchema:
    properties:
      currency:
        example: UAH
        type: string
      value:
        example: "1250.50"
        type: string
    type: object
  schemas.CreateAccountSchema:
    properties:
      colorHex:
        example: '#1E88E5'
        maxLength: 7
        type: string
      currency:
        example: UAH
        type: string
      defaultValue:
        example: "1250.50"
        type: string
      description:
        example: Main salary card
        maxLength: 256
        type: string
      name:
        example: Monobank card
        maxLength: 64
        type: string
    required:
    - colorHex
    - currency
    - name
    type: object
  schemas.HealthSchema:
    properties:
      status:
        example: ok
        type: string
    type: object
  schemas.PaginatedResponse-schemas_AccountSchema:
    properties:
      items:
        items:
          $ref: '#/definitions/schemas.AccountSchema'
        type: array
      page:
        type: integer
      take:
        type: integer
      total:
        type: integer
    type: object
  schemas.ReadinessSchema:
    properties:
      database:
        example: ok
        type: string
      error:
        type: string
      status:
        example: ok
        type: string
    type: object
  schemas.UpdateUserSchema:
    properties:
      currency:
        example: UAH
        type: string
      language:
        example: uk
        type: string
    type: object
  schemas.UserSchema:
    properties:
      currency:
        example: UAH
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      language:
        example: uk
        type: string
      name:
        example: John Doe
        type: string
      pictureUrl:
        example: https://example.com/picture.png
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Personal finance API for accounts, transactions and budgets
  title: Kopiika API
  version: "1.0"
paths:
  /health:
    get:
      description: Reports that the process is serving requests. Checks no dependencies.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.HealthSchema'
      summary: Health check
      tags:
      - health
  /health/live:
    get:
      description: Reports that the process is alive. Always 200 while the server
        is up, so a failure means the process should be restarted.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.HealthSchema'
      summary: Liveness probe
      tags:
      - health
  /health/ready:
    get:
      description: Reports whether the service can serve traffic, verifying the database
        connection. Returns 503 when a dependency is unavailable.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ReadinessSchema'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/schemas.ReadinessSchema'
      summary: Readiness probe
      tags:
      - health
  /v1/accounts:
    get:
      consumes:
      - application/json
      description: List the authenticated user's accounts, oldest first
      parameters:
      - default: 1
        description: Page number, 1-indexed
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PaginatedResponse-schemas_AccountSchema'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new account for the authenticated user
      parameters:
      - description: Account to create
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateAccountSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.AccountSchema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create account
      tags:
      - accounts
  /v1/accounts/{accountId}:
    delete:
      consumes:
      - application/json
      description: Soft delete one of the authenticated user's accounts. Idempotent.
      parameters:
      - description: Account id
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Get one of the authenticated user's accounts by id
      parameters:
      - description: Account id
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AccountSchema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account
      tags:
      - accounts
  /v1/users:
    post:
      consumes:
      - application/json
      description: Create the local configuration for the authenticated Cognito user.
        Idempotent.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserSchema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Setup user
      tags:
      - users
  /v1/users/me:
    get:
      consumes:
      - application/json
      description: Get the currently authenticated user's configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserSchema'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's language and default currency
      parameters:
      - description: User configuration
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateUserSchema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UserSchema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"